requirements.txt
aiohttp==3.6.2
aiodns==2.0.0
asyncio==3.4.3
async-timeout==3.0.1
attrs==19.1.0
chardet==3.0.4
discord==1.0.1
discord.py==1.7.3
idna==2.8
multidict==4.5.2
websockets==6.0
yarl==1.3.0
pytz
pillow==8.0.1
